connection established between players

+++ game.js
@@ -15,51 +15,52 @@ exports.initialize = function(server) {
   io = socket(server);
 
   io.on('connection', function(socket) {
-    console.log('user connected');
 
     socket.on('join', function(gameId){
       const players = ( games[gameId] || 0 ) + 1;
-      console.log('players', players);
 
       if( players <= 2){
         socket.join(gameId);
-        console.log('player has now joined ', socket.room);
         games[gameId] = players;
         socket.gameId = gameId;
       }
 
       if( players === 2){
-        // socket.broadcast.emit('joined',  )
-          db.Game.findById( gameId ).populate('player1', 'player2').exec(function(game){
-            console.log(game);
-            io.to( gameId ).emit('log', 'hey')
-            io.to( gameId ).emit('start game', game );
-          })
+        db.Game.findById( gameId ).populate('player1','player2').exec(function(err, game){
 
+          io.to( gameId ).emit('start game', game );
+        })
       }
     })
 
 
 
     socket.on('disconnect', function(){
-      // update list of users in chat, client-side
-      // io.sockets.emit('updateusers', usernames);
-      // echo globally that this client has left
-      socket.broadcast.emit('playerError');
-      delete games[ socket.gameId ];
-      // socket.leave(socket.room);
 
-      db.Game.findByIdAndRemove( socket.gameId ).exec(function(err, deletedItem){
-        if( err ) return console.log(err);
-        console.log('deleted')
-        // io.sockets.clients( socket.gameId ).forEach(function(s){
-        //   s.leave( socket.gameId );
-        // });
-      })
+      console.log('disconnected ', socket.gameId);
+
+      io.in( socket.gameId ).emit('playerError');
+      if( games[ socket.gameId ] ){
+        // if game hasnt been deleted already
+        delete games[ socket.gameId ];
+        // socket.leave(socket.room);
+
+        db.Game.findByIdAndRemove( socket.gameId ).exec(function(err, deletedItem){
+          if( err ) return console.log(err);
+          // console.log(io.sockets.clients( socket.gameId ))
+          // io.sockets.clients( socket.gameId ).forEach(function(s){
+          //   s.leave( socket.gameId );
+          // });
+        })
+      }
 
     });
 
-    // socket.emit('chess message', 'hey loser')
+    socket.on('move', function(){
+      console.log('hey')
+      // socket.brodcast
+
+    })
   });
 
   return io;
